+++ src/id_translation/fetching/_sql_fetcher.py
@@ -228,10 +228,10 @@ class SqlFetcher(AbstractFetcher[str, IdType]):
             query = str(select.compile(self.engine, compile_kwargs={"literal_binds": True}))
             if len(query) > query_length_limit:
                 query = query[:query_length_limit]
-            self.logger.debug(f"Full SELECT-query using {self.engine}:\n{query}", extra=logger_extra)
+            self.logger.debug(f"Full SELECT-query using {self._estr}:\n{query}", extra=logger_extra)
         except Exception as e:
             self.logger.debug(
-                f"Failed to render full SELECT-query using {self.engine}:\n{select}",
+                f"Failed to render full SELECT-query using {self._estr}:\n{select}",
                 exc_info=e,
                 extra=logger_extra,
             )